style(feat_ext): split `protein.py` script
Split into `protein_edges.py` and `protein_nodes.py` for maintainability.
